requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
Bottleneck==1.3.2
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.9.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsons==1.4.2
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kt-legacy==1.0.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
missingno==0.4.2
multimethod==1.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandas-profiling==3.0.0
pandocfilters==1.4.3
parso==0.8.2
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
ppscore==1.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
pydot==1.4.2
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
seaborn==0.11.1
six==1.15.0
smart-open==5.1.0
smmap==4.0.0
SQLAlchemy==1.4.19
squarify==0.4.3
streamlit==0.82.0
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typish==1.9.2
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
visions==0.7.1
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
